OrderService/requirements.txt
bcrypt==4.0.1
email-validator==2.0.0.post2
fastapi==0.95.1
httpcore==0.17.2
httpx==0.24.1
passlib==1.7.4
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
regex==2022.10.31
uvicorn==0.22.0
sqlalchemy==1.4.31
psycopg2-binary==2.9.6